% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cnes.R
\name{search_cnes}
\alias{search_cnes}
\title{Consulta de Estabelecimentos de Saúde no Site do Ministério da Saúde}
\usage{
search_cnes(nu_cnes)
}
\arguments{
\item{nu_cnes}{Um vetor de códigos CNES.}
}
\value{
A função retorna um dataframe com informações básicas dos estabelecimentos de saúde.
}
\description{
\code{search_cnes} é uma função de busca de informações básicas dos estabelecimentos de saúde do Brasil.
}
\examples{
library(sustools)
dados <- search_cnes(nu_cnes)

}
\references{
Site do Ministério da Saúde para consulta de estabelecimentos de saúde.\url{http://cnes.datasus.gov.br/pages/estabelecimentos/consulta.jsp}
}
\author{
Ronaldo Fernandes Santos Alves. Fundação Oswaldo Cruz, Instituto de Comunicação e Informação Científica e Tecnológica em Saúde.
}
