% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_childdeath.R
\name{stop_childdeath}
\alias{stop_childdeath}
\title{Codificação de Causas de Óbitos Evitáveis na Infância}
\usage{
stop_childdeath(dados, var_nome)
}
\arguments{
\item{dados}{Um data frame contendo os dados de interesse.}

\item{var_nome}{Um vetor de caracteres contendo os códigos da CID-10.}
}
\value{
A função retorna um data frame enriquecido com a variável "stop_code".
}
\description{
\code{stop_childdeath} é uma função de codificação de causas de óbitos evitáveis em crianças menores de 5 anos de idade. A função cria um vetor de códigos identificadores (stop_code), cuja estrutura compreende variados níveis hierárquicos de classificação de causas de óbitos evitáveis.
}
\examples{
library(sustools)
dados <- stop_childdeath(dados, dados$var_nome)

}
\references{
Malta DC, Duarte EC, Almeida MF, et al. Lista de causas de mortes evitáveis por intervenções do Sistema Único de Saúde do Brasil. Epidemiologia e Serviços de Saúde, 16(4):233-244, out-dez 2007.\url{http://dx.doi.org/10.5123/S1679-49742007000400002}

Malta DC, Sardinha LMV, Moura L, et al. Atualização da lista de causas de mortes evitáveis por intervenções do Sistema Único de Saúde do Brasil. Epidemiologia e Serviços de Saúde, 19(2):173-176, abr-jun 2010.\url{http://dx.doi.org/10.5123/S1679-49742010000200010}
}
\author{
Ronaldo Fernandes Santos Alves. Fundação Oswaldo Cruz, Instituto de Comunicação e Informação Científica e Tecnológica em Saúde.
}
