% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_hospitalization.R
\name{stop_hospitalization}
\alias{stop_hospitalization}
\title{Codificação de Codificações Sensíveis à Atenção Primária em Saúde}
\usage{
stop_hospitalization(dados, var_nome)
}
\arguments{
\item{dados}{Um data frame contendo os dados de interesse.}

\item{var_nome}{Um vetor de caracteres contendo os códigos da CID-10.}
}
\value{
A função retorna um data frame enriquecido com a variável "stop_code".
}
\description{
\code{stop_hospitalization} é uma função de classificação de códigos CID-10 segundo a lista brasileira de condições sensíveis à atenção primária em saúde. Esta função cria um vetor de códigos identificadores (stop_code), cuja estrutura compreende os diversos níveis da classificação de condições sensíveis à atenção primária em saúde.
}
\examples{
library(sustools)
dados <- stop_hospitalization(dados, dados$var_nome)

}
\references{
Alfradique ME, Bonolo PF, Dourado I, et al. Internações por condições sensíveis à atenção primária: a construção da lista brasileira como ferramenta para medir o desempenho do sistema de saúde (Projeto ICSAP - Brasil). Cadernos de Saúde Pública, 25(6), jun 2009.\url{https://doi.org/10.1590/S0102-311X2009000600016}
}
\author{
Ronaldo Fernandes Santos Alves. Fundação Oswaldo Cruz, Instituto de Comunicação e Informação Científica e Tecnológica em Saúde.
}
